+++ RELEASES.md
@@ -7,6 +7,7 @@ Released on TBD (UTC).
 - Added `BacktestDataIterator` to backtest engine to provide on-the-fly data loading (#2545), thanks @faysou
 - Added support for `MarkPriceUpdate` streaming from catalog (#2582), thanks @bartolootrit
 - Added `activation_price` support for trailing stop orders (#2610), thanks @hope2see
+- Added `raise_exception` config option for `BacktestRunConfig` (default `False` to retain current behavior) which will raise exceptions to interrupt a nodes run process
 - Added `UnixNanos::is_zero()` convenience method to check for a zero/epoch value
 - Introduce HyperSync client to blockchain adapter (#2606), thanks @filipmacek